# swagger.yaml
swagger: "2.0"
info:
  title: Bibus Reservas API
  description: >
    API RESTful para acceder al sistema de consulta de salas de la
    **Biblioteca de la Universidad de Sevilla**.


    _Esta API **no es oficial** ni ha sido desarrollada por ningún miembro de
    la Biblioteca de la US. No se ofrece ningún tipo de garantía en cuanto a
    su funcionamiento._


    **El autor no se responsabiliza del mal uso que se le pueda dar**.


    Más información en https://bigomby.github.io/bibing-salas-api/
  version: 0.6.0
  contact:
    email: bigomby@gmail.com
host: api.salas.gonebe.com
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json

tags:
  - name: Salas
    description: Consulta de salas
  - name: Reservas
    description: Reserva de salas

paths:
  /salas:
    x-swagger-router-controller: salas
    get:
      tags: [Salas]
      summary: Consultar salas
      description: Devuelve información sore las salas de la Biblioteca
      operationId: getAll
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/Room"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
  /salas/{roomID}:
    x-swagger-router-controller: salas
    get:
      tags: [Salas]
      summary: Consultar una sala
      description: Devuelve información sore una sala de la Biblioteca
      operationId: get
      parameters:
        - name: roomID
          in: path
          description: Room number
          required: true
          type: integer
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Room"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/Error"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
  /reservas/{date}:
    x-swagger-router-controller: reservas
    get:
      tags: [Reservas]
      summary: Consultar reserva
      description: Obtiene el ID de una reserva dada una fecha
      operationId: getReservation
      parameters:
        - name: date
          in: path
          description: Fecha de la reserva
          required: true
          type: string
        - name: username
          in: query
          description: Usuario para realizar autenticación
          required: true
          type: string
        - name: password
          in: query
          description: Contraseña para realizar autenticación
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ReservationResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ReservationError"
  /reservas:
    x-swagger-router-controller: reservas
    post:
      tags: [Reservas]
      summary: Realizar reserva
      description: Reserva una sala de la biblioteca
      operationId: reserve
      parameters:
        - name: reservation
          in: body
          description: Datos de la reserva
          required: true
          schema:
            $ref: "#/definitions/ReservationRequest"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ReservationResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
    delete:
      tags: [Reservas]
      summary: Cancelar reserva
      description: Cancela una reserva de una sala de la biblioteca
      operationId: cancel
      parameters:
        - name: cancellation
          in: body
          description: Datos de la reserva
          required: true
          schema:
            $ref: "#/definitions/CancelRequest"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/CancelResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"

definitions:
  Room:
    required:
      - id
      - turns
    properties:
      id:
        type: integer
      turns:
        type: array
        items:
          $ref: "#/definitions/Turn"
  Turn:
    required:
      - id
      - time
      - available
    properties:
      id:
        type: integer
      time:
        type: string
      available:
        type: boolean
  ReservationResponse:
    required:
      - room
      - turn
      - reservation
      - date
      - time
    properties:
      room:
        type: integer
      turn:
        type: integer
      reservation:
        type: string
      date:
        type: string
      time:
        type: string
  ReservationRequest:
      required:
        - name
        - uvus
        - email
        - room
        - turn
        - date
        - username
        - password
      properties:
        name:
          description: Nombre del alumno. Los espacios deben sustituirse por signos '+'
          type: string
        uvus:
          description: Uvus del alumno
          type: string
        email:
          description: Correo para recibir la notificación
          type: string
        room:
          description: Número de la sala
          type: integer
        turn:
          description: Número del turno (Entre 1 y 6)
          type: integer
        date:
          description: Fecha en formato 'dd-mm-aaaa'
          type: string
        username:
          description: Uvus del alumno usado para la autenticación
          type: string
        password:
          description: Contrasela correspondiente al uvus.
          type: string
  CancelRequest:
      required:
        - reservation
      properties:
        reservation:
          description: ID de la reserva
          type: string
  CancelResponse:
      required:
        - reservation
      properties:
        reservation:
          description: ID de la reserva
          type: string
  Error:
    required:
      - message
    properties:
      message:
        type: string
  ReservationError:
    required:
      - message
    properties:
      message:
        type: string
      reservation:
        $ref: "#/definitions/ReservationResponse"